Re-generated cookie-cutter template
